requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.3
django-bootstrap4==1.1.1
djangorestframework==3.12.2
google-api-python-client==1.5.5
httplib2==0.19.0
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
uritemplate==3.0.1
